feat: add dispatch_git_command IPC handler and regenerate bindings (GIT-021)
Add #[tauri::command] #[specta::specta] dispatch_git_command to
commands/shell.rs, register it in lib.rs invoke_handler and specta type
collection, and regenerate tauri-bindings.ts. Fix usize -> u32 in
StashEntryDto and GitCommandDto StashPop/StashDrop for specta BigInt
compatibility.

+++ src-tauri/crates/tabby-contracts/src/git_dtos.rs
@@ -102,7 +102,7 @@ pub struct BlameEntryDto {
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct StashEntryDto {
-    pub index: usize,
+    pub index: u32,
     pub message: String,
     pub date: String,
 }
@@ -198,14 +198,14 @@ pub enum GitCommandDto {
     },
     StashPop {
         pane_id: String,
-        index: Option<usize>,
+        index: Option<u32>,
     },
     StashList {
         pane_id: String,
     },
     StashDrop {
         pane_id: String,
-        index: usize,
+        index: u32,
     },
     DiscardChanges {
         pane_id: String,
